@@ -1,14 +1,14 @@
 <?php
 
+use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Kernel;
 use Symfony\Component\Config\Loader\LoaderInterface;
-
 
 class AppKernel extends Kernel
 {
     public function registerBundles()
     {
-        $bundles = array(
+        $bundles = [
             new Symfony\Bundle\FrameworkBundle\FrameworkBundle(),
             new Symfony\Bundle\SecurityBundle\SecurityBundle(),
             new Symfony\Bundle\TwigBundle\TwigBundle(),
@@ -18,29 +18,49 @@ class AppKernel extends Kernel
             new Sensio\Bundle\FrameworkExtraBundle\SensioFrameworkExtraBundle(),
             new AppBundle\AppBundle(),
             new Naturalheritage\SearchBundle\NaturalheritageSearchBundle(),
-	    #ftheeten 2017 10 24
-	    //new FOS\ElasticaBundle\FOSElasticaBundle(),
-	    //new Symfony\Component\Serializer\Serializer(),
-	    //new JMS\SerializerBundle\JMSSerializerBundle(),
-	    new ONGR\ElasticsearchBundle\ONGRElasticsearchBundle(),
-	    new Braincrafted\Bundle\BootstrapBundle\BraincraftedBootstrapBundle(),
- 	    new Tetranz\Select2EntityBundle\TetranzSelect2EntityBundle(),
-	    new Symfony\Bundle\AsseticBundle\AsseticBundle(),
-//new Elasticsearch\ClientBuilder();
-        );
+            new ONGR\ElasticsearchBundle\ONGRElasticsearchBundle(),
+            new Naoned\OaiPmhServerBundle\NaonedOaiPmhServerBundle(),
+            new Naturalheritage\OaipmhBundle\NaturalheritageOaipmhBundle(),
+			//new Naturalheritage\OaipmhBundle\OaiPmhRulerNh(),
+        ];
 
-        if (in_array($this->getEnvironment(), array('dev', 'test'), true)) {
+        if (in_array($this->getEnvironment(), ['dev', 'test'], true)) {
             $bundles[] = new Symfony\Bundle\DebugBundle\DebugBundle();
             $bundles[] = new Symfony\Bundle\WebProfilerBundle\WebProfilerBundle();
             $bundles[] = new Sensio\Bundle\DistributionBundle\SensioDistributionBundle();
-            $bundles[] = new Sensio\Bundle\GeneratorBundle\SensioGeneratorBundle();
+
+            if ('dev' === $this->getEnvironment()) {
+                $bundles[] = new Sensio\Bundle\GeneratorBundle\SensioGeneratorBundle();
+                $bundles[] = new Symfony\Bundle\WebServerBundle\WebServerBundle();
+            }
         }
 
         return $bundles;
     }
 
+    public function getRootDir()
+    {
+        return __DIR__;
+    }
+
+    public function getCacheDir()
+    {
+        return dirname(__DIR__).'/var/cache/'.$this->getEnvironment();
+    }
+
+    public function getLogDir()
+    {
+        return dirname(__DIR__).'/var/logs';
+    }
+
     public function registerContainerConfiguration(LoaderInterface $loader)
     {
+        $loader->load(function (ContainerBuilder $container) {
+            $container->setParameter('container.autowiring.strict_mode', true);
+            $container->setParameter('container.dumper.inline_class_loader', true);
+
+            $container->addObjectResource($this);
+        });
         $loader->load($this->getRootDir().'/config/config_'.$this->getEnvironment().'.yml');
     }
 }
